requirements.txt
llama-index
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
# llama-index-postprocessor-flag-embedding-reranker
# git+https://github.com/FlagOpen/FlagEmbedding.git
llama-parse
python-dotenv